[refactor] bnv set
week1과 week2,3 과정을 한번에 표기하고자 수정함
