+++ app/src/main/java/org/android/go/sopt/account/GalleryFragment.kt
@@ -1,4 +1,4 @@
-package org.android.go.sopt.playlist
+package org.android.go.sopt.account
 
 import android.os.Bundle
 import android.view.LayoutInflater
